pyproject.toml
[project]
name = "lovelaice"
version = "0.3.6"
description = "An AI-powered assistant for your terminal and editor."
readme = "Readme.md"
requires-python = ">=3.10"
dependencies = [
    "googlesearch-python>=1.3.0",
    "openai>=1.58.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
]

[project.scripts]
lovelaice = "lovelaice:cli.run"

[project.optional-dependencies]
api = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "pytest>=8.3.4",
]
